requirements.txt
# local package
cryptography
fastembed
langchain[llms]
langchain_anthropic
langchain-openai
langchain-community
tiktoken
cachetools
beautifulsoup4
python-dotenv>=0.5.1
qdrant-client
sentence-transformers
# external requirements
